Split tests in clearly style
